refactor: put helper methods into helper file for recipe repository
