+++ lib/cubit/recipe_cubits/add_recipe_cubit.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:dotd/cubit/recipe_cubits/recipes_cubit.dart';
 import 'package:dotd/data/models/recipe_model/recipe_model.dart';
-import 'package:dotd/data/repositories/recipe_repository.dart';
+import 'package:dotd/data/repositories/recipe_repository/recipe_repository.dart';
 import 'package:meta/meta.dart';
 
 part 'add_recipe_state.dart';
