+++ lib/cubit/recipe_cubits/recipes_cubit.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:dotd/data/models/recipe_model/recipe_model.dart';
-import 'package:dotd/data/repositories/recipe_repository.dart';
+import 'package:dotd/data/repositories/recipe_repository/recipe_repository.dart';
 import 'package:meta/meta.dart';
 
 part 'recipes_state.dart';
