+++ lib/presentation/router.dart
@@ -6,7 +6,7 @@ import 'package:dotd/data/models/screen_arguments.dart';
 import 'package:dotd/data/network_services/analytics_services.dart';
 import 'package:dotd/data/network_services/recipe_drift_database/recipe_drift_database.dart';
 import 'package:dotd/data/network_services/recipe_network_service.dart';
-import 'package:dotd/data/repositories/recipe_repository.dart';
+import 'package:dotd/data/repositories/recipe_repository/recipe_repository.dart';
 import 'package:dotd/presentation/screens/add_recipe_screen.dart';
 import 'package:dotd/presentation/screens/details_recipe_screen.dart';
 import 'package:dotd/presentation/screens/edit_recipe_screen.dart';
